Code divided into classes
