+++ Frinder/data/shopping_list.txt
@@ -1,2 +1,2 @@
-butter
-sugar
+cheese
+milk
