+++ Frinder/data/user_ingredients.txt
@@ -1,2 +1,1 @@
-milk
 eggs
